:sparkles: feat: GamePulse 설정 클래스를 Demo.Web으로 이동 및 네임스페이스 수정
