+++ NET10/Demo.Web/Configs/JwtConfig.cs
@@ -1,4 +1,4 @@
-namespace GamePulse.Configs;
+namespace Demo.Web.Configs;
 
 /// <summary>
 /// Jwt 인증
