+++ NET10/GamePulse/GamePulse.csproj
@@ -46,6 +46,7 @@
   <ItemGroup>
     <ProjectReference Include="..\Demo.Application\Demo.Application.csproj" />
     <ProjectReference Include="..\Demo.Infra\Demo.Infra.csproj" />
+    <ProjectReference Include="..\Demo.Web\Demo.Web.csproj" />
   </ItemGroup>
 
 </Project>
